refactor first_place component
