+++ src/task/methods/first_place/config.vsh.yaml
@@ -4,12 +4,15 @@ functionality:
   name: first_place
   info:
     label: First Place
-    summary: "FILL IN: A one sentence summary of this method."
+    summary: An ensemble of LSTM, GRU, and 1D CNN models
     description: |
-      FILL IN: A (multi-line) description of how this method works.
-    reference: bibtex_reference_key
-    documentation_url: https://url.to/the/documentation
-    repository_url: https://github.com/organisation/repository
+      An ensemble of LSTM, GRU, and 1D CNN models with a variety of input features derived from ChemBERTa embeddings,
+      one-hot encoding of cell type/small molecule pairs, and various statistical measures of target gene expression.
+      The models were trained with a combination of MSE, MAE, LogCosh, and BCE loss functions to improve their
+      robustness and predictive performance. The approach also included data augmentation techniques to ensure
+      generalization and account for noise in the data.
+    documentation_url: https://www.kaggle.com/competitions/open-problems-single-cell-perturbations/discussion/459258
+    repository_url: https://github.com/Jean-KOUAGOU/1st-place-solution-single-cell-pbs/tree/main
 
   arguments:
     - name: "--extra_output"
@@ -40,17 +43,34 @@ functionality:
   resources:
     - type: python_script
       path: script.py
-    - path: helper_script.py
+    - path: helper_classes.py
+    - path: helper_functions.py
+    - path: models.py
+    - path: predict.py
+    - path: prepare_data.py
+    - path: train.py
     
 platforms:
   - type: docker
-    image: ghcr.io/openproblems-bio/base_python:1.0.4
+    image: ghcr.io/openproblems-bio/base_pytorch_nvidia:1.0.4
     run_args: ["--shm-size=2g"]
     setup:
       - type: python
-        packages: [ fastparquet, pandas==2.1.3, pyarrow==14.0.1, tqdm==4.66.1, scikit-learn==1.3.2, torch==2.1.1, transformers==4.35.2, matplotlib==3.8.2 ]
+        packages:
+          - torch==2.1.1
+          - --index-url
+          - https://download.pytorch.org/whl/cu121
+      - type: python
+        packages:
+          - fastparquet
+          - pandas==2.1.3
+          - pyarrow==14.0.1
+          - tqdm==4.66.1
+          - scikit-learn==1.3.2
+          - transformers==4.35.2
+          - matplotlib==3.8.2
 
   - type: native
   - type: nextflow
     directives:
-      label: [midtime,midmem,highcpu,midsharedmem,gpu]
+      label: [hightime,highmem,highcpu,midsharedmem,gpu]
